orchestration/airflow/requirements.txt
aiokafka==0.10.0
avro==1.11.3
black==24.2.0
confluent-kafka==2.3.0
coverage==7.3.2
defusedxml==0.7.1
flake8==7.0.0
isort==5.13.2
moto==5.0.1
neo4j==5.18.0
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.0
pytest-mock==3.12.0
pytz==2024.1
requests==2.31.0
structlog==24.1.0
urllib3==1.26.18